requirements.txt
torch
torchaudio
librosa
numpy
scikit-learn
matplotlib
pyyaml
tqdm
soundfile
